template/python3-flask-debian-tpu/function/requirements.txt
Pillow
shapely
numpy
